contract update

+++ smart-contracts/contracts/lib/MerkleProof.sol
@@ -1,9 +1,9 @@
 pragma solidity >=0.6.0;
 
 library MerkleProof {
-    function keccak256MerkleProof(bytes32[8] memory proof, uint8 path, bytes32 leaf) internal pure returns(bytes32) {
+    function keccak256MerkleProof(bytes32[8] memory proof, uint path, bytes32 leaf) internal pure returns(bytes32) {
         bytes32 root = leaf;
-        for(uint8 i=0; i<8; i++) {
+        for(uint i = 0; i < 8; i++) {
             root = path >> i & 1 == 0 ?  keccak256(abi.encode(leaf, proof[i])) : keccak256(abi.encode(proof[i], leaf));
         }
         return root;
